change Bayes Factor for posterior probability
git-svn-id: svn://svn.r-forge.r-project.org/svnroot/ebs@40 edb9625f-4e0d-4859-8d74-9fd3b1da38cb

+++ pkg/src/LogPoisson.cpp
@@ -26,7 +26,7 @@ double LogPoisson::operator()(int a, int b)
   int S = LesObs.SumInSegment(a,b);
   double L = LesObs.LogFactorialInSegment(a,b);
   int n = b-a;
-  double Res = lgamma(S+alpha)-(S+alpha)*log(n+beta)-L+alpha*log(beta)-lgamma(alpha);
+  double Res = lgamma(S+alpha)-(S+alpha)*log((double)(n)+beta)-L+alpha*log(beta)-lgamma(alpha);
   return Res;
 }
